fritzcollectd: Fix password handling, implement LAN statistics
The action WANCommonInterfaceConfig, GetAddonInfos can be used only
without authentication. Otherwise it raises "KeyError: 'GetAddonInfos".

In order to access services/actions that are only available with
authentication, we create a second FritzConnection object.

Report LAN statistics (if a password is configured) as an example for
data that can be read only with authentication.

Fixes #6, fixes #9