fritzcollectd: Recover from XMLSyntaxError in read callback
In rare cases reading from the FRITZ!Box can trigger XMLSyntaxError
exceptions (for example: 'Opening and ending tag mismatch: HR line 1
and BODY, line 1, column 180 (line 1)'. If this happens the connection
seems to be unusable and the error would happen also on subsequent
calls. Reinitialize the connection if this happens.

This has been reported for a FRITZ!Box 6490 (6.52) when the device is
remotely reset by Unitymedia.

See #12.