fritzcollectd: Adapt to fritzconnection 0.8
The fritzconnection library has changed the way the services are named.
Calls that work without authentication now use corrcet short names such
as WANIPConn:1 or WANCommonIFC:1. This fixes the root cause of
requiring two connections to the fritzbox in fritzcollectd (see the
comments in this commit's diff).

Changing to a single connection will be done in a separate commit.

See: https://bitbucket.org/kbr/fritzconnection/issues/37/unknown-service-wancommoninterfaceconfig-1
Fixes: https://github.com/fetzerch/fritzcollectd/issues/31